requirements.txt
# Core dependencies
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Optional GPU acceleration
cupy-cuda11x>=12.0.0  # Optional
# triton>=2.0.0  # Optional - Linux only

# Deployment
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
onnxruntime>=1.16.0

# Visualization
plotly>=5.17.0
dash>=2.14.0
dash-bootstrap-components>=1.4.2
kaleido>=0.2.0

# Utils
pyyaml>=6.0
tqdm>=4.66.0
pandas>=2.0.0
py-cpuinfo>=9.0.0
optuna>=3.0.0
#GPUtil>=1.4.0 #For GPU
psutil>=5.9.0
# numba>=0.57.0  # for CPU optimization


# Testing
pytest>=7.4.0
pytest-cov>=4.1.0